LocalScanner: Read the stored results for all packages at once
Refactor the `LocalScanner` to use the previously introduced function to
read scan results for multiple packages at once from the scan results
storage to benefit from potential performance gains.

The new implementation does not use coroutines anymore to parallelize
the reading of scan results and scanning of packages. One reason is that
the now used `ScanResultsStorage.read()` function is responsible for
implementing the optimal way of reading multiple scan results. The other
reason is that downloading of scan results is usually a lot faster than
scanning packages, therefore the performance benefit of downloading
results and scanning in parallel was negligible.

This change is also a preparation for an upcoming change in the
`PostgresStorage`. It will be changed to store scan results by VCS
repository instead of by `Package` which will require to group packages
by VCS repository before starting to download scan results. At the same
time the `LocalScanner` needs to stay compatible with other storages
that store results by `Package`. This will be easier to implement with
the changes of this commit.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>