AbstractStorageFunTest: Assert some actions in place
In the functions that test reading scan results, assert the success of
adding the results to the storage in place. This slightly violates the
AAA pattern, but makes the code a lot shorter and more readable, and
puts the focus of the last section on the assertions of the read results
which are actually under test.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>